Transfer interface from domain layer Rename Interface, Namespace
