+++ NotificationMicroservice.Application/Abstractions/IMessageApplicationService.cs
@@ -1,7 +1,6 @@
 ﻿using NotificationMicroservice.Application.Model.Message;
-using NotificationMicroservice.Domain.Interfaces.Services;
 
-namespace NotificationMicroservice.Application.Interface
+namespace NotificationMicroservice.Application.Abstractions
 {
     public interface IMessageApplicationService : IService<MessageModel, CreateMessageModel, Guid>
     {
