+++ NotificationMicroservice.Application/Services/MessageService.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
-using NotificationMicroservice.Application.Interface;
+using NotificationMicroservice.Application.Abstractions;
 using NotificationMicroservice.Application.Model.Message;
-using NotificationMicroservice.DataAccess.Interfaces;
+using NotificationMicroservice.DataAccess.Abstractions;
 using NotificationMicroservice.Domain.Entities;
 
 namespace NotificationMicroservice.Application.Services
