+++ NotificationMicroservice/Controllers/MessageTypeController.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
-using NotificationMicroservice.Application.Interface;
+using NotificationMicroservice.Application.Abstractions;
 using NotificationMicroservice.Application.Model.Type;
 using NotificationMicroservice.Contracts.Type;
